Dragonball: update dependencies
Since rust-vmm and dragonball-sandbox has introduced several updates
such as vPMU support for aarch64, we also need to update Dragonball
dependencies to include those changes.

Update:
virtio-queue to v0.6.0
kvm-ioctls to v0.12.0
dbs-upcall to v0.2.0
dbs-virtio-devices to v0.2.0
kvm-bindings to v0.6.0

Also, several aarch64 features are updated because of dependencies
changes:
1. update mpidr parameters.
2. update vcpu hotplug API.
3. update vpmu related API.
4. adjust unit test cases for aarch64 Dragonball.

fixes: #6268

+++ src/dragonball/src/vcpu/aarch64.rs
@@ -112,7 +112,7 @@ impl Vcpu {
             .map_err(VcpuError::REGSConfiguration)?;
         }
 
-        self.mpidr = regs::read_mpidr(&self.fd).map_err(VcpuError::REGSConfiguration)?;
+        self.mpidr = regs::read_mpidr(&self.fd).map_err(VcpuError::REGSConfiguration)? as u64;
 
         Ok(())
     }
